# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Çoklu Firma Muhasebe ve Cari Takip Paneli API
servers:
  - url: /api
    description: API temel yolu
...
paths:
...
  /starlink-planlari:
    get:
      operationId: listStarlinkPlanlari
      tags: [starlink-planlari]
      summary: Starlink planlarını listele
      parameters:
        - name: sirketId
          in: query
          schema:
            type: integer
        - name: gemiId
          in: query
          schema:
            type: integer
        - name: aktif
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Starlink planı listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StarlinkPlani"
    post:
      operationId: createStarlinkPlani
      tags: [starlink-planlari]
      summary: Yeni Starlink planı oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StarlinkPlaniInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StarlinkPlani"
  /starlink-planlari/{id}:
    get:
      operationId: getStarlinkPlani
      tags: [starlink-planlari]
      summary: Starlink planını getir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Starlink planı
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StarlinkPlani"
    patch:
      operationId: updateStarlinkPlani
      tags: [starlink-planlari]
      summary: Starlink planını güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StarlinkPlaniUpdate"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StarlinkPlani"
    delete:
      operationId: deleteStarlinkPlani
      tags: [starlink-planlari]
      summary: Starlink planını sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi

  # ── Ekipmanlar ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    StarlinkPlani:
      type: object
      required: [id, sirketId, cariId, gemiId, planAdi, baslangicTarihi, bitisTarihi, aylikUcret, paraBirimi, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        sirketAd:
          type: ["string", "null"]
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        gemiId:
          type: integer
        gemiAd:
          type: ["string", "null"]
        planAdi:
          type: string
        hizMbps:
          type: ["integer", "null"]
        baslangicTarihi:
          type: string
          format: date
        bitisTarihi:
          type: string
          format: date
        aylikUcret:
          type: number
        paraBirimi:
          type: string
        otomatikYenileme:
          type: boolean
        aktif:
          type: boolean
        notlar:
          type: ["string", "null"]
        kalanGun:
          type: ["integer", "null"]
        olusturmaTarihi:
          type: string
          format: date-time
    StarlinkPlaniInput:
      type: object
      required: [sirketId, cariId, gemiId, planAdi, baslangicTarihi, bitisTarihi, aylikUcret, paraBirimi]
      properties:
        sirketId:
          type: integer
        cariId:
          type: integer
        gemiId:
          type: integer
        planAdi:
          type: string
        hizMbps:
          type: integer
        baslangicTarihi:
          type: string
          format: date
        bitisTarihi:
          type: string
          format: date
        aylikUcret:
          type: number
        paraBirimi:
          type: string
        otomatikYenileme:
          type: boolean
        aktif:
          type: boolean
        notlar:
          type: string
    StarlinkPlaniUpdate:
      type: object
      properties:
        planAdi:
          type: string
        hizMbps:
          type: integer
        bitisTarihi:
          type: string
          format: date
        aylikUcret:
          type: number
        otomatikYenileme:
          type: boolean
        aktif:
          type: boolean
        notlar:
          type: string

    # ── Ekipman ───────────────────────────────────────────────────────────
...